Review (#720): guard the setter, separate TENSOR from SYM_TENSOR, light up the tests
Three findings from adversarial review, all real.

**The shape guard was in the wrong place.** It lived only in __init__, but a
solver reassigns DFDt.psi_fn = flux.T on every setup (six call sites in
solvers.py), so the guard was absent from the one path that is actually
driven. A larger matrix silently TRUNCATED -- the component writer reads
psi_fn[i, j] for the slots it already has -- and a smaller one died later with
"IndexError: Index out of range: a[1]". The check now lives in the setter.

**SYM_TENSOR and TENSOR share a symbolic shape** and differ only in storage
width (3 against 4 in 2-D), so the shape check could not separate them: a full
tensor handed to a symmetric history passed and failed later with a bare
broadcast error naming neither vtype. When psi_fn is a variable it knows its
own width, so compare that.

**A non-symmetric psi_fn under SYM_TENSOR loses its lower entries**, and the
two implementations disagree about which triangle survives. On
[[1+x, 2+y], [100.0, 3+x*y]]:

    nodal SemiLagrangian        -> [[1.4499, 100.0 ], [100.0 , 3.2137]]
    IntegrationPointSemiLagrangian -> [[1.4603, 2.4667], [2.4667, 3.2148]]

Neither averages and neither warned. This class now warns and names the
entries it drops; the nodal divergence is pre-existing, so it carries a
TODO(BUG) with the measurement rather than a silent change (charter section 9).

**The tests never ran.** test_006[2-9] and test_0070 matched no batch glob in
scripts/test.sh -- ten files, including the whole integration-point suite
(0064-0067), swarm repopulation, mid-time velocity and the VE stress history.
They are not covered by the disabled test_06*py line either; that one is
0600-0699. Verified all 83 passing, then wired the range in.

The tensor tests move into tests/test_0066_integration_point_slcn.py: they are
SLCN-at-the-integration-points tests, that file is their family, and the file
they were in duplicated the number of the existing
test_0068_swarm_repopulation.py.

Full level_1 and tier_a: 1202 passed, 3 skipped, 1 xfailed.
